Fix wrong resize anchor point
Previously, when dragging the left border of a window with the mouse, there
was a bug where it would snap the top level surface's geometry X coordinate
directly to the position of the mouse, as if you started the resize right on the border. This also affected the other (right, bottom, and top) borders.

I think that the previous resize code was hard to understand. Honestly I
have not spent a lot of time trying to understand why t didn't work and
I wrote another resize algorithm instead. Now instead of working directly
with widths and heights which are complicated we work with the borders (left,
right, top and bottom). This is easier to understand IMO.